refactor: use standard `build`, `test`, `build-for-testing` commands
