+++ .github/workflows/sdk.ai.yml
@@ -45,12 +45,14 @@ jobs:
           - os: macos-15
             target: iOS
             xcode: Xcode_26.2
-            test_filter_arg: --build-only
-            log_prefix: xcodebuild-build-only
+            command: build-for-testing
+            test_filter_arg: ""
+            log_prefix: xcodebuild-build-for-testing
           - os: macos-26
             target: iOS
             # TODO: Bump to 26.4 when "Unable to find a destination matching the provided destination specifier" is resolved.
             xcode: Xcode_26.2
+            command: test
             test_filter_arg: --exclude IntegrationTests-SPM/LiveSessionTests
             log_prefix: xcodebuild
     runs-on: ${{ matrix.os }}
@@ -69,7 +71,7 @@ jobs:
         path: .build
         key: ${{ needs.spm.outputs.cache_key }}
     - name: Run integration tests
-      run: scripts/repo.sh tests run --secrets ./scripts/secrets/AI.json --platforms ${{ matrix.target }} --xcode ${{ matrix.xcode }} ${{ matrix.test_filter_arg }} AI
+      run: scripts/repo.sh ${{ matrix.command }} --secrets ./scripts/secrets/AI.json --platforms ${{ matrix.target }} --xcode ${{ matrix.xcode }} ${{ matrix.test_filter_arg }} AI
     - name: Upload xcodebuild logs
       if: failure()
       uses: actions/upload-artifact@b7c566a772e6b6bfb58ed0dc250532a479d7789f # v6.0.0
@@ -97,7 +99,7 @@ jobs:
         path: .build
         key: ${{ needs.spm.outputs.cache_key }}
     - name: Run live integration tests
-      run: scripts/repo.sh tests run --secrets ./scripts/secrets/AI.json --platforms iOS --xcode Xcode_26.2 --filter IntegrationTests-SPM/LiveSessionTests AI
+      run: scripts/repo.sh test --secrets ./scripts/secrets/AI.json --platforms iOS --xcode Xcode_26.2 --filter IntegrationTests-SPM/LiveSessionTests AI
     - name: Upload xcodebuild logs
       if: failure()
       uses: actions/upload-artifact@b7c566a772e6b6bfb58ed0dc250532a479d7789f # v6.0.0
