+++ scripts/repo/README.md
@@ -4,10 +4,14 @@ This project includes commands that are too long and complicated to properly
 maintain in a bash script, or that have unique option/flag constraints that
 are better represented in a swift project.
 
-## Tests
+## Available Commands
 
-Commands for interacting with integration tests in the repo.
+- `build`: Build the integration test app for a given SDK.
+- `build-for-testing`: Build the integration test app and tests without running.
+- `test`: Run integration tests for a given SDK.
+- `decrypt`: Decrypt secret files for a test run.
 
 ```sh
-./scripts/repo.sh tests --help
+./scripts/repo.sh --help
+./scripts/repo.sh test --help
 ```
